requirements-main.txt
# Integrating dbt-core
dbt-core
dbt-snowflake # Installing dbt-snowflake adapter (https://docs.getdbt.com/docs/core/connect-data-platform/snowflake-setup)

# Integrating library to interact with AWS S3
boto3

# Integrating Snowflake via Jupyter Notebook
snowflake-connector-python

# Pandas for data validation / observability scripts
pandas

# Meltano core for orchestration and ingestion
meltano

# Stripe for API ingestion
stripe


# Integrating SLQ Engines
# sqlalchemy # Because I run sql commands on a python script and because it depends on psycopg2
# psycopg2 # A PostgreSQL adapter for Python required to connect to PostgreSQL databases using SQLAlchemy
# openpyxl # Python library for reading and writing Excel
# xlrd # Python library for reading and writing Excel

# Integrating Software Engienering Best Practices Tools
# pycodestyle # A tool to check your Python code against some of the style conventions in PEP 8
# pytest # A framework that makes it easy to write small, readable tests

# Integration Other Tools
# faker # To ingest fake data via the python script with sql
# tweepy # If we want to access twitter api and manipulate jsons
python-dotenv # Reads key-value pairs from a .env file and can set them as environment variables
python-dateutil # For date manipulation
requests # Requests for API ingestion (Allows you to send HTTP/1.1 requests extremely easily)

# Slack notifications
slack_sdk

# Great Expectations for advanced tests
great_expectations

# Data Analytics Tools
# seaborn # For visualization
# matplotlib # For visualization

# Integrating Jupyter Notebooks
ipykernel
jupyter